refine bot testing and ac_member* handling
